requirements.txt
# ============================================================
# Общие зависимости
# ============================================================
numpy>=1.21.0
requests>=2.31.0

# ============================================================
# Анализ тональности текста (Hugging Face)
# ============================================================
transformers>=4.35.0
torch>=2.0.0

# ============================================================
# Распознавание речи (Wav2Vec2 + librosa)
# ============================================================
librosa>=0.10.0
soundfile>=0.12.0

# ============================================================
# Классификация изображений (ResNet-50 + torchvision)
# ============================================================
torchvision>=0.15.0
Pillow>=10.0.0

# ============================================================
# Классификация видео (VideoMAE + decord)
# ============================================================
decord>=0.6.0

# ============================================================
# FastAPI для моделей (Sentiment + LLM)
# ============================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0

# ============================================================
# Тестирование (pytest, асинхронные тесты)
# ============================================================
pytest==7.4.3
pytest-asyncio==0.21.1

# ============================================================
# Рекомендательные системы
# ============================================================
scikit-learn>=1.3.0
pandas>=2.0.0
scipy>=1.10.0
# mlxtend>=0.22.0   # опционально для market basket analysis